+++ projects/angular-auth-oidc-client/src/lib/auth-state/check-auth.service-mock.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
-import { LoginResponse } from './login/login-response';
+import { LoginResponse } from '../login/login-response';
 
 @Injectable()
 export class CheckAuthServiceMock {
