+++ projects/angular-auth-oidc-client/src/lib/auth-state/check-auth.service.spec.ts
@@ -1,32 +1,32 @@
 import { TestBed, waitForAsync } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { of, throwError } from 'rxjs';
-import { AuthStateService } from './auth-state/auth-state.service';
-import { AuthStateServiceMock } from './auth-state/auth-state.service-mock';
-import { AutoLoginService } from './auto-login/auto-login.service';
-import { CallbackService } from './callback/callback.service';
-import { CallbackServiceMock } from './callback/callback.service-mock';
-import { PeriodicallyTokenCheckService } from './callback/periodically-token-check.service';
-import { PeriodicallyTokenCheckServiceMock } from './callback/periodically-token-check.service-mock';
-import { RefreshSessionService } from './callback/refresh-session.service';
-import { RefreshSessionServiceMock } from './callback/refresh-session.service.mock';
+import { AutoLoginService } from '../auto-login/auto-login.service';
+import { CallbackService } from '../callback/callback.service';
+import { CallbackServiceMock } from '../callback/callback.service-mock';
+import { PeriodicallyTokenCheckService } from '../callback/periodically-token-check.service';
+import { PeriodicallyTokenCheckServiceMock } from '../callback/periodically-token-check.service-mock';
+import { RefreshSessionService } from '../callback/refresh-session.service';
+import { RefreshSessionServiceMock } from '../callback/refresh-session.service.mock';
+import { StsConfigLoader } from '../config/loader/config-loader';
+import { StsConfigLoaderMock } from '../config/loader/config-loader-mock';
+import { CheckSessionService } from '../iframe/check-session.service';
+import { CheckSessionServiceMock } from '../iframe/check-session.service-mock';
+import { SilentRenewService } from '../iframe/silent-renew.service';
+import { SilentRenewServiceMock } from '../iframe/silent-renew.service-mock';
+import { LoggerService } from '../logging/logger.service';
+import { LoggerServiceMock } from '../logging/logger.service-mock';
+import { PopUpService } from '../login/popup/popup.service';
+import { PopUpServiceMock } from '../login/popup/popup.service-mock';
+import { StoragePersistenceService } from '../storage/storage-persistence.service';
+import { StoragePersistenceServiceMock } from '../storage/storage-persistence.service-mock';
+import { UserServiceMock } from '../user-data/user-service-mock';
+import { UserService } from '../user-data/user.service';
+import { CurrentUrlService } from '../utils/url/current-url.service';
+import { CurrentUrlServiceMock } from '../utils/url/current-url.service-mock';
+import { AuthStateService } from './auth-state.service';
+import { AuthStateServiceMock } from './auth-state.service-mock';
 import { CheckAuthService } from './check-auth.service';
-import { StsConfigLoader } from './config/loader/config-loader';
-import { StsConfigLoaderMock } from './config/loader/config-loader-mock';
-import { CheckSessionService } from './iframe/check-session.service';
-import { CheckSessionServiceMock } from './iframe/check-session.service-mock';
-import { SilentRenewService } from './iframe/silent-renew.service';
-import { SilentRenewServiceMock } from './iframe/silent-renew.service-mock';
-import { LoggerService } from './logging/logger.service';
-import { LoggerServiceMock } from './logging/logger.service-mock';
-import { PopUpService } from './login/popup/popup.service';
-import { PopUpServiceMock } from './login/popup/popup.service-mock';
-import { StoragePersistenceService } from './storage/storage-persistence.service';
-import { StoragePersistenceServiceMock } from './storage/storage-persistence.service-mock';
-import { UserServiceMock } from './user-data/user-service-mock';
-import { UserService } from './user-data/user.service';
-import { CurrentUrlService } from './utils/url/current-url.service';
-import { CurrentUrlServiceMock } from './utils/url/current-url.service-mock';
 
 describe('CheckAuthService', () => {
   let checkAuthService: CheckAuthService;
