+++ projects/angular-auth-oidc-client/src/lib/auto-login/auto-login-all-routes.guard.spec.ts
@@ -2,8 +2,8 @@ import { TestBed, waitForAsync } from '@angular/core/testing';
 import { RouterStateSnapshot, UrlSegment } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Observable, of } from 'rxjs';
-import { CheckAuthService } from '../check-auth.service';
-import { CheckAuthServiceMock } from '../check-auth.service-mock';
+import { CheckAuthService } from '../auth-state/check-auth.service';
+import { CheckAuthServiceMock } from '../auth-state/check-auth.service-mock';
 import { ConfigurationService } from '../config/config.service';
 import { ConfigurationServiceMock } from '../config/config.service.mock';
 import { LoginService } from '../login/login.service';
