+++ projects/angular-auth-oidc-client/src/lib/auto-login/auto-login-partial-routes.guard.spec.ts
@@ -4,8 +4,8 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { of } from 'rxjs';
 import { AuthStateService } from '../auth-state/auth-state.service';
 import { AuthStateServiceMock } from '../auth-state/auth-state.service-mock';
-import { CheckAuthService } from '../check-auth.service';
-import { CheckAuthServiceMock } from '../check-auth.service-mock';
+import { CheckAuthService } from '../auth-state/check-auth.service';
+import { CheckAuthServiceMock } from '../auth-state/check-auth.service-mock';
 import { ConfigurationService } from '../config/config.service';
 import { ConfigurationServiceMock } from '../config/config.service.mock';
 import { LoginService } from '../login/login.service';
