+++ projects/angular-auth-oidc-client/src/lib/login/popup/popup-login.service.spec.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { TestBed, waitForAsync } from '@angular/core/testing';
 import { of } from 'rxjs';
-import { CheckAuthService } from '../../check-auth.service';
-import { CheckAuthServiceMock } from '../../check-auth.service-mock';
+import { CheckAuthService } from '../../auth-state/check-auth.service';
+import { CheckAuthServiceMock } from '../../auth-state/check-auth.service-mock';
 import { AuthWellKnownService } from '../../config/auth-well-known/auth-well-known.service';
 import { AuthWellKnownServiceMock } from '../../config/auth-well-known/auth-well-known.service-mock';
 import { LoggerService } from '../../logging/logger.service';
