+++ projects/angular-auth-oidc-client/src/lib/oidc.security.service.ts
@@ -3,9 +3,9 @@ import { map, Observable, switchMap } from 'rxjs';
 import { AuthOptions } from './auth-options';
 import { AuthenticatedResult } from './auth-state/auth-result';
 import { AuthStateService } from './auth-state/auth-state.service';
+import { CheckAuthService } from './auth-state/check-auth.service';
 import { CallbackService } from './callback/callback.service';
 import { RefreshSessionService } from './callback/refresh-session.service';
-import { CheckAuthService } from './check-auth.service';
 import { AuthWellKnownEndpoints } from './config/auth-well-known/auth-well-known-endpoints';
 import { AuthWellKnownService } from './config/auth-well-known/auth-well-known.service';
 import { ConfigurationService } from './config/config.service';
